+++ packages/sdk/src/parser/service/tailordb/index.ts
@@ -1,4 +1,8 @@
 export { stringifyFunction, tailorUserMap } from "./field";
 export { parseTypes } from "./type-parser";
-export { TailorDBServiceConfigSchema, TailorDBTypeSchema } from "./schema";
+export {
+  GQL_PERMISSION_INVALID_OPERAND_MESSAGE,
+  TailorDBServiceConfigSchema,
+  TailorDBTypeSchema,
+} from "./schema";
 export type * from "./types";